requirements.txt
google-genai
pandas
google-api-python-client
gspread
oauth2client
isodate
pytz
numpy
requests
python-dotenv
